requirements.txt
autopep8==2.0.2
beautifulsoup4==4.12.2
bidict==0.22.1
certifi==2023.7.22
charset-normalizer==3.2.0
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.27.2
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.4
deep-sort-realtime==1.3.2
dnspython==2.4.2
docstring-parser==0.15
eventlet==0.33.3
filelock==3.12.3
fonttools==4.42.1
ftfy==6.1.1
gdown==4.7.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
lap==0.4.0
lapx==0.5.5
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
mypy==1.5.1
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
psutil==5.9.5
py-cpuinfo==9.0.0
pycodestyle==2.11.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
scipy==1.9.3
seaborn==0.12.2
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
supervision==0.14.0
sympy==1.12
torch==2.0.1
torchreid==0.2.5
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typed-argument-parser==1.8.1
types-Pillow==10.0.0.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.164
urllib3==2.0.4
wcwidth==0.2.8
websockets==12.0
wsproto==1.2.0
